requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pyyaml>=6.0
ccxt>=4.0.0
python-binance>=1.0.19

# Database
sqlalchemy>=2.0.0
pymysql>=1.1.0

# Web framework
flask>=3.0.0
flask-cors>=4.0.0

# Logging
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Utilities
requests>=2.31.0
python-dateutil>=2.8.2
